+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+# 3.2.0
+
+- Bump iOS version from 3.5.9 to 3.8.9
+- Bump android version from 3.5.9 to 3.8.12
+- remove `setAdIdEnabled`
+- rename `sendEvent` to `logConversion` and add `revenue` param
+- add `trackVerified`
+- add `trackVerifiedToken`
+- add `isUsingRemoteTrackingOptions`
+- update `autocompleteQuery` with param add `expandUnits`
+- add `validateAddress`
+- add `App Attest` to ios example
+- add `Play Integrity API` to android example
+- update example project
+
 # 3.1.7
 
 - Update ios event channel